create components for each graph
