+++ client/components/Home.jsx
@@ -12,7 +12,7 @@ export default class Home extends Component {
 
   handleUpload = data => {
     store.dispatch(uploadData(data))
-    history.push('/graph')
+    history.push('/graph-dataset')
   }
 
   render() {
