+++ client/components/graphs/index.js
@@ -1,0 +1,6 @@
+export {default as LineChartGraph} from './LineChartGraph'
+export {default as BarChartGraph} from './BarChartGraph'
+export {default as AreaChartGraph} from './AreaChartGraph'
+export {default as RadarChartGraph} from './RadarChartGraph'
+export {default as ScatterChartGraph} from './ScatterChartGraph'
+export {default as PieChartGraph} from './PieChartGraph'
