+++ client/components/index.js
@@ -9,4 +9,4 @@ export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as Home} from './Home'
 export {default as SearchBar} from './SearchBar'
-export {default as Graph} from './Graph'
+export {default as GraphDataset} from './GraphDataset'
