+++ client/routes.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup, UserHome, Home, SearchBar, Graph} from './components'
+import {Login, Signup, UserHome, Home, SearchBar, GraphDataset} from './components'
 import {me} from './store'
 
 /**
@@ -23,7 +23,7 @@ class Routes extends Component {
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
         <Route path="/search" component={SearchBar} />
-        <Route path="/graph" component={Graph} />
+        <Route path="/graph-dataset" component={GraphDataset} />
         <Route path="/" component ={Home}/>
         {isLoggedIn && (
           <Switch>
